+++ vjer/build.py
@@ -125,4 +125,4 @@ def build() -> None:
     """This is the main entry point."""
     VjerAction('build', cast(VjerStep, BuildStep)).execute()
 
-# cSpell:ignore batcave fileutil pythonval buildargs vjer
+# cSpell:ignore pythonval buildargs